Extract PillSheetAppearanceMode to other file
